Factor out a more abstract WatcherThread from EventWorkMonitor to re-use
